chore(test): add hidden test_bus page for Arraial/BUS testing
